Add game() function
Add a new function called 'game' that calls the existing 'playRound'
function to play a 5 round game, keeping score and reporting the winner
or loser at the end. The purpose of this refactor is to organize the
code and make it more reusable, allowing the game to be easily played
multiple times.